Add more setu for tests.
